+++ core/app-version.js
@@ -1,1 +1,1 @@
-export const APP_VERSION = '20260630-8';
+export const APP_VERSION = '20260630-9';