+++ modules/graph-assumption-builder.js
@@ -1,10 +1,10 @@
-import { GraphUtils } from './graph-utils.js?v=20260630-8';
+import { GraphUtils } from './graph-utils.js?v=20260630-9';
 import {
     buildAssumptionNodeTooltip,
     buildAttackEdgeTooltip,
     buildJunctionTooltip,
     buildTopNodeTooltip
-} from './graph-tooltip-builder.js?v=20260630-8';
+} from './graph-tooltip-builder.js?v=20260630-9';
 
 // Cap on the number of minimal support sets enumerated per contrary, guarding
 // against combinatorial blow-up on pathological (deeply disjunctive) frameworks.